intrinsicArray: Don't use sp for unused [skip ci]

+++ src/runtime/lfortran_intrinsic_array.f90
@@ -1,5 +1,4 @@
 module lfortran_intrinsic_array
-use, intrinsic :: iso_fortran_env, only: sp => real32, dp => real64
 implicit none
 
 interface
@@ -64,20 +63,20 @@ contains
 
 ! int --------------------------------------------------------------------------
 
-elemental integer(sp) function iint(x)
+elemental integer function iint(x)
 integer, intent(in) :: x
 end function
 
-elemental integer(sp) function rint(x)
+elemental integer function rint(x)
 real, intent(in) :: x
 end function
 
-elemental integer(sp) function iintk(x, kind)
+elemental integer function iintk(x, kind)
 integer, intent(in) :: x(:)
 integer, intent(in) :: kind
 end function
 
-elemental integer(sp) function rintk(x, kind)
+elemental integer function rintk(x, kind)
 real, intent(in) :: x(:)
 integer, intent(in) :: kind
 end function
